requirements.txt
dnspython
termcolor
openpyxl